FIX: Update do-not-disturb icon in real-time on glimmer header
To achieve this, a new ReactiveTargetDate 'resource' is introduced. This is constructed using a callback, which should return the target date (in this case, when do-not-disturb will end). The utility then provides an autotrackable `.hasPassed` getter which returns a boolean.

This resource needs to be properly torn down at the end of the application lifecycle in tests, so the logic is moved out of the user model and into a new Notifications service.